Honor optimization preference for planet gears
The module planetary_gear has a parameter to specify whether the gears should be optimized ("create holes for
material/weight reduction resp. surface increase, if geometry allows"). However, this parameter was only
affecting the sun gear, due to the fact that the parameter was not being passed when creating
the planet gears. This commit adds the parameter wherever it was missing.